requirements.txt
django
djangorestframework
psycopg2-binary
Pillow
drf-yasg